Implemented user info editing & cleaned up room booking code & UI.

+++ LibCalBooker/Controllers/BookingsController.cs
@@ -56,43 +56,30 @@ namespace LibCalBooker.Controllers
         public async Task<IActionResult> BookRoom(TimeSlot timeSlot)
         {
             ApplicationUser user = await _userManager.GetUserAsync(User);
-            return RedirectToAction(nameof(Index));
+            await LibCalSession.BookRoom(timeSlot, user);
+			return RedirectToAction(nameof(Index));
         }
 
 		// GET: Bookings
 		[Authorize]
         public async Task<IActionResult> Index()
         {
-            var libCalContext = _context.Bookings.Include(b => b.Room);
-            return View(await libCalContext.ToListAsync());
+            ApplicationUser user = await _userManager.GetUserAsync(User);
+            if (user == null)
+			{
+				return RedirectToAction("Identity/Account/Logout");
+			}
+			var bookings = _context.Bookings.Where(b => b.BookerID == user.Id);
+            return View(await bookings.ToListAsync());
         }
 
-        // GET: Bookings/Details/5
-        [Authorize]
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var booking = await _context.Bookings
-                .Include(b => b.Room)
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (booking == null)
-            {
-                return NotFound();
-            }
-
-            return View(booking);
-        }
 
         // GET: Bookings/Create
         [Authorize]
         public IActionResult Create()
         {
             TimeSpan start = new TimeSpan(7, 0, 0);
-            List<String> times = new List<String>();
+            List<string> times = new List<string>();
 
 			while (start != TimeSpan.FromHours(21)+TimeSpan.FromMinutes(45))
             {
@@ -112,10 +99,6 @@ namespace LibCalBooker.Controllers
         [Authorize]
         public async Task<IActionResult> Create([Bind("Id,BookingDate,BookingTime,RoomID")] Booking booking)
         {
-			//         if (_context.Bookings.Where(b => booking.RoomId == b.RoomId && (booking.BookingTime - b.BookingTime).Hours < 4).Any())
-			//{
-			//             ModelState.AddModelError("BookingTime", "Room is already booked from: " + booking.BookingTime.ToString());
-			//}
 			booking.BookerID = (await _userManager.GetUserAsync(User)).Id;
 			if (ModelState.IsValid || ModelState.ErrorCount == 1)
             {
@@ -128,7 +111,7 @@ namespace LibCalBooker.Controllers
                 var balls = ModelState.ErrorCount;
             }
 			TimeSpan start = new TimeSpan(7, 0, 0);
-			List<String> times = new List<String>();
+			List<string> times = new();
 
 			while (start.Hours < 21 && start.Minutes < 45)
 			{
@@ -140,60 +123,7 @@ namespace LibCalBooker.Controllers
             return View(booking);
         }
 
-        // GET: Bookings/Edit/5
-        [Authorize]
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
 
-            var booking = await _context.Bookings.FindAsync(id);
-            if (booking == null)
-            {
-                return NotFound();
-            }
-            ViewData["RoomId"] = new SelectList(_context.Rooms, "Id", "Id", booking.RoomID);
-            return View(booking);
-        }
-
-        // POST: Bookings/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        [Authorize]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,FirstName,LastName,RegistrationNumber,UCardNumber,BookerEmail,SecondaryEmail,BookingTime,RoomId")] Booking booking)
-        {
-            if (id != booking.Id)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(booking);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!BookingExists(booking.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            ViewData["RoomId"] = new SelectList(_context.Rooms, "Id", "Id", booking.RoomID);
-            return View(booking);
-        }
 
         // GET: Bookings/Delete/5
         [Authorize]
@@ -222,7 +152,11 @@ namespace LibCalBooker.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var booking = await _context.Bookings.FindAsync(id);
-            if (booking != null)
+            if (booking.BookerID != (await _userManager.GetUserAsync(User)).Id)
+			{
+				return Unauthorized();
+			}
+			if (booking != null)
             {
                 _context.Bookings.Remove(booking);
             }
